fix: break budget dependency cycle
