+++ runtime/src/budget/evaluator.ts
@@ -2,48 +2,7 @@ import type Database from "bun:sqlite";
 
 import { ensureBudgetCapWindow, getBudgetWork, listBudgetCaps } from "../db/budget-limits.js";
 import { getDb } from "../db/connection.js";
-import type { BudgetCap, BudgetWorkStatus } from "./types.js";
-
-export type BudgetDecisionAction = "allow" | "warn" | "pause" | "stop";
-
-export interface BudgetProviderEvidence {
-  capId: string;
-  providerId: string;
-  accountRef: string;
-  quotaDimension: string;
-  windowId: string;
-  fetchedAt: string;
-  resetsAt: string | null;
-  stale: boolean;
-  availability: string;
-  valueMicros: number | null;
-  source: string;
-}
-
-export interface BudgetBlocker {
-  capId: string;
-  capRevision: number;
-  scope: BudgetCap["scope"];
-  metric: BudgetCap["metric"];
-  windowId: string;
-  limit: number;
-  allowance: number;
-  knownUsage: number;
-  remaining: number | null;
-  unknownEvents: number;
-  reason: "exhausted" | "unknown_pricing" | "missing_provider_evidence" | "stale_provider_evidence" | "provider_account_mismatch" | "provider_window_expired";
-  detail: string;
-}
-
-export interface BudgetDecision {
-  action: BudgetDecisionAction;
-  workId: string;
-  checkedAt: string;
-  blockers: BudgetBlocker[];
-  warnings: BudgetBlocker[];
-  applicableCapIds: string[];
-  warningsOnly: boolean;
-}
+import type { BudgetBlocker, BudgetCap, BudgetDecision, BudgetDecisionAction, BudgetProviderEvidence, BudgetWorkStatus } from "./types.js";
 
 type UsageTotal = { known_usage: number; unknown_events: number };
